Revert "fix(GraphLoader): type mismatch for vertex and edge"
This reverts commit 005992d092212121eb8901d7e7e9f7862f801802.

+++ src/main/scala/com/omnia/spark/benchmarks/graphLoader/ParquetGraphLoader.scala
@@ -13,7 +13,7 @@ class ParquetGraphLoader extends LogTrait {
     val pRDD = p.rdd
     step("[ParquetGraphLoader]Parquet => RDD[Row]")
 
-    val eRDD = pRDD.map(row => Edge[Int](row.getInt(0), row.getInt(1)))
+    val eRDD = pRDD.map(row => Edge[Long](row.getLong(0), row.getLong(1)))
     step("[ParquetGraphLoader]RDD[Row] => RDD[Edge]")
 
     val e = EdgeRDD.fromEdges(eRDD)
